fixes gcc warnings (-Wzero-as-null-pointer-constant and -Wuseless-cast)

+++ capture/chunked_image_transport.hpp
@@ -52,7 +52,7 @@ private:
     {
         size_t total_len = 0;
         while (total_len < len){
-            if (static_cast<size_t>(stream.end - stream.p) >= static_cast<size_t>(len - total_len)){
+            if (static_cast<size_t>(stream.end - stream.p) >= (len - total_len)){
                 stream.in_copy_bytes(*pbuffer + total_len, len - total_len);
                 *pbuffer += len;
                 return;
